Adds an analytics script to record web vitals to Vercel Analytics
Eleventy does not have good support for building JS, so
we have to use a Javascript *template* to compile entrypoint
assets and, using the render() function, transform them with
esbuild into an actual file in the build/ directory.

The built file can then be used in templates.

The JS included here is assigning a global function to window.webVitals.
If I execute the code in the script itself, I can't find a way
to access the analyticsId, so I chose to load the script and invoke
it in separate steps. The invocation in the Nunjucks template
can then pass the analyticsId, since the Nunjucks template _can_
access the global data file.